rework: override only the CJS wrapper args, not module.filename
Move the '[worker eval]' / '.' substitution into evaluateCommonJSModuleOnce
so only the __filename/__dirname arguments passed to the CJS wrapper are
affected. module.filename and the requireMap key stay on the blob URL,
which keeps resolve() and require error messages behaving exactly as
before and avoids the requireMap key/filename mismatch.

Move the test to its own file so the gate can run it in isolation.
