+++ test/js/node/worker_threads/worker_threads.test.ts
@@ -295,7 +295,7 @@ test("support require in eval for a file that doesnt exist", async () => {
     worker.on("message", resolve);
     worker.on("error", resolve);
   });
-  expect(result.toString()).toInclude(`error: Cannot find module './fixture-invalid.js' from '[worker eval]'`);
+  expect(result.toString()).toInclude(`error: Cannot find module './fixture-invalid.js' from 'blob:`);
   await worker.terminate();
 });
 
@@ -308,24 +308,6 @@ test("support worker eval that throws", async () => {
   expect(result.toString()).toInclude("Unexpected throw");
   expect(result.name).toBe("SyntaxError");
   await worker.terminate();
-});
-
-test("eval worker __filename/__dirname match node's '[worker eval]' / '.'", async () => {
-  const code = `
-    const { parentPort } = require("node:worker_threads");
-    parentPort.postMessage({ f: __filename, d: __dirname, a1: process.argv[1] });
-  `;
-  const run = () =>
-    new Promise((resolve, reject) => {
-      const w = new Worker(code, { eval: true });
-      w.on("message", m => {
-        w.terminate().then(() => resolve(m), reject);
-      });
-      w.on("error", reject);
-    });
-  const [r1, r2] = await Promise.all([run(), run()]);
-  expect(r1).toEqual({ f: "[worker eval]", d: ".", a1: "[worker eval]" });
-  expect(r2).toEqual(r1);
 });
 
 describe("execArgv option", async () => {
